Move useGetOnChainCancellation to own file
