+++ src/cow-react/common/hooks/useCancelOrder/useSendOnChainCancellation.ts
@@ -1,12 +1,9 @@
 import { useCallback } from 'react'
-
-import { useEthFlowContract, useGP2SettlementContract } from 'hooks/useContract'
 import { Order } from 'state/orders/actions'
 import { useRequestOrderCancellation, useSetOrderCancellationHash } from 'state/orders/hooks'
 import { useTransactionAdder } from 'state/enhancedTransactions/hooks'
 import { useWalletInfo } from '@cow/modules/wallet'
-import { getIsEthFlowOrder } from '@cow/modules/swap/containers/EthFlowStepper'
-import { getEthFlowCancellation, getOnChainCancellation, OnChainCancellation } from './onChainCancellation'
+import { useGetOnChainCancellation } from './useGetOnChainCancellation'
 
 export function useSendOnChainCancellation() {
   const { chainId } = useWalletInfo()
@@ -32,23 +29,5 @@ export function useSendOnChainCancellation() {
       }
     },
     [addTransaction, getOnChainCancellation, cancelPendingOrder, chainId, setOrderCancellationHash]
-  )
-}
-
-export function useGetOnChainCancellation(): (order: Order) => Promise<OnChainCancellation> {
-  const ethFlowContract = useEthFlowContract()
-  const settlementContract = useGP2SettlementContract()
-
-  return useCallback(
-    (order: Order) => {
-      const isEthFlowOrder = getIsEthFlowOrder(order)
-
-      if (isEthFlowOrder) {
-        return getEthFlowCancellation(ethFlowContract!, order)
-      } else {
-        return getOnChainCancellation(settlementContract!, order)
-      }
-    },
-    [ethFlowContract, settlementContract]
   )
 }
